move wall, floor and player, classes to separate file
